refactor(eo): move out of oidc folder
